Split player1 movement & timer
