+++ Assets/Scripts/Players/PlayerOneMovement.cs
@@ -3,9 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class PlayerOne : MonoBehaviour {
-    // test comment
-
+public class PlayerOneMovement : MonoBehaviour {
     //camera
     [SerializeField] private CameraOneRotator cam;
     private int state = 1;
@@ -26,12 +24,6 @@ public class PlayerOne : MonoBehaviour {
         return state;
     }
 
-    //timer and death
-    [SerializeField] private float time = 500f;
-    private bool lose;
-    public Text TimerText;
-
-
     private Vector3 tempVelocity;
     private bool isGrounded;
 	private Rigidbody rb;
@@ -40,8 +32,6 @@ public class PlayerOne : MonoBehaviour {
 
 	void Start() {
 		rb = GetComponent<Rigidbody> ();
-        lose = false;
-        SetTimerText();
         inputAxis = gameManager.GetInputAxis();
     }
 
@@ -123,24 +113,5 @@ public class PlayerOne : MonoBehaviour {
         {
             rb.velocity += Vector3.up * Physics.gravity.y * (lowerJumpMultiplier - 1) * Time.deltaTime;
         }
-
-        //timer countdown
-        time -= Time.deltaTime;
-        SetTimerText();
-        if(time < 0)
-        {
-            lose = true;
-            GameOver();
-        }
-    }
-
-    void SetTimerText()
-    {
-        TimerText.text = "Time: " + time.ToString("F0");
-    }
-
-    public bool GameOver()
-    {
-        return lose;
     }
 }
