+++ Assets/Scripts/Timer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 66f93ba2ae77ace4388bbc5723b2216e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
